+++ seqlog/__init__.py
@@ -16,10 +16,10 @@ from seqlog.structured_logging import set_callback_on_failure as _set_callback_o
 
 __author__ = 'Adam Friedman'
 __email__ = 'tintoy@tintoy.io'
-__version__ = '0.3.27'
+__version__ = '0.3.30'
 
 
-def configure_from_file(file_name, override_root_logger=True, support_extra_properties=False, support_stack_info=False):
+def configure_from_file(file_name, override_root_logger=True, support_extra_properties=False, support_stack_info=False, ignore_seq_submission_errors=False):
     """
     Configure Seq logging using YAML-format configuration file.
 
@@ -33,10 +33,13 @@ def configure_from_file(file_name, override_root_logger=True, support_extra_prop
     :type support_extra_properties: bool
     :param support_stack_info: Support attaching of stack-trace information (if available) to log records?
     :type support_stack_info: bool
+    :param ignore_seq_submission_errors: Ignore errors encountered while sending log records to Seq?
+    :type ignore_seq_submission_errors: bool
     """
 
     configure_feature(FeatureFlag.EXTRA_PROPERTIES, support_extra_properties)
     configure_feature(FeatureFlag.STACK_INFO, support_stack_info)
+    configure_feature(FeatureFlag.IGNORE_SEQ_SUBMISSION_ERRORS, ignore_seq_submission_errors)
 
     with open(file_name) as config_file:
         config = yaml.load(config_file, Loader=yaml.SafeLoader)
@@ -44,7 +47,7 @@ def configure_from_file(file_name, override_root_logger=True, support_extra_prop
     configure_from_dict(config, override_root_logger)
 
 
-def configure_from_dict(config, override_root_logger=True, use_structured_logger=True, support_extra_properties=False, support_stack_info=False):
+def configure_from_dict(config, override_root_logger=True, use_structured_logger=True, support_extra_properties=False, support_stack_info=False, ignore_seq_submission_errors=False):
     """
     Configure Seq logging using a dictionary.
 
@@ -60,10 +63,13 @@ def configure_from_dict(config, override_root_logger=True, use_structured_logger
     :type support_extra_properties: bool
     :param support_stack_info: Support attaching of stack-trace information (if available) to log records?
     :type support_stack_info: bool
+    :param ignore_seq_submission_errors: Ignore errors encountered while sending log records to Seq?
+    :type ignore_seq_submission_errors: bool
     """
 
     configure_feature(FeatureFlag.EXTRA_PROPERTIES, support_extra_properties)
     configure_feature(FeatureFlag.STACK_INFO, support_stack_info)
+    configure_feature(FeatureFlag.IGNORE_SEQ_SUBMISSION_ERRORS, ignore_seq_submission_errors)
 
     if override_root_logger:
         _override_root_logger()
@@ -79,7 +85,9 @@ def log_to_seq(server_url, api_key=None, level=logging.WARNING,
                name='DefaultSeqLogger', batch_size=10, auto_flush_timeout=None,
                additional_handlers=None, override_root_logger=False,
                json_encoder_class=None, support_extra_properties=False,
-               support_stack_info=False,  handler=None, formatter=None):
+               support_stack_info=False,
+               ignore_seq_submission_errors=False,
+               **kwargs):
     """
     Configure the logging system to send log entries to Seq.
 
@@ -102,6 +110,8 @@ def log_to_seq(server_url, api_key=None, level=logging.WARNING,
     :type support_extra_properties: bool
     :param support_stack_info: Support attaching of stack-trace information (if available) to log records?
     :type support_stack_info: bool
+    :param ignore_seq_submission_errors: Ignore errors encountered while sending log records to Seq?
+    :type ignore_seq_submission_errors: bool
     :param handler: a custom/user-provided handler class for the StructuredLogger class
         If not specified, SeqLogHandler will be initialized and used.
     :type logging.Handler(class)
@@ -114,7 +124,10 @@ def log_to_seq(server_url, api_key=None, level=logging.WARNING,
 
     configure_feature(FeatureFlag.EXTRA_PROPERTIES, support_extra_properties)
     configure_feature(FeatureFlag.STACK_INFO, support_stack_info)
-    
+    configure_feature(FeatureFlag.IGNORE_SEQ_SUBMISSION_ERRORS, ignore_seq_submission_errors)
+
+    logging.setLoggerClass(StructuredLogger)
+
     if override_root_logger:
         _override_root_logger()
 
@@ -210,7 +223,7 @@ def set_global_log_properties(**properties):
     Configure the properties to be added to all structured log entries.
 
     :param properties: Keyword arguments representing the properties.
-    :type properties: dict
+    :type properties: str
     """
 
     _set_global_log_properties(**properties)